pyproject.toml
[tool.poetry]
name = "zereAgent-Defi"
version = "0.1.0"
#package-mode = false
description = "A launch-pad for AI agents"
authors = ["Eric <eric2612115@gmail.com>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
openai = "^1.57.2"
prompt-toolkit = "^3.0.48"
anthropic = "^0.42.0"
farcaster = "^0.7.11"
goat-sdk = "^0.1.1"
goat-sdk-wallet-web3 = "^0.1.1"
web3 = "^6.20.3"
goat-sdk-plugin-erc20 = "^0.1.0"
goat-sdk-plugin-coingecko = "^0.1.1"
solders = "^0.21.0,<0.24.0"
aiohttp = "^3.11.11"
requests = "2.32.3"
requests-oauthlib = "^1.3.1"
together = "^1.3.14"
motor = "^3.7.0"
uvicorn = "^0.34.0"
google-genai = "^1.5.0"
fastapi = "^0.115.11"

[tool.poetry.extras]
server = ["fastapi", "uvicorn", "requests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
